FEAT: Add licenses to the top of relevant files. DOCS: readme table formatted correctly as per MD060 CLEANUP: Move default metadata classes so they live with their appropriate interfaces.
