+++ FiftyOne.Did.Cloud/FlowElements/DidCloudEngine.cs
@@ -1,4 +1,26 @@
-﻿using FiftyOne.Did.Core.Data;
+﻿/* *********************************************************************
+ * This Original Work is copyright of 51 Degrees Mobile Experts Limited.
+ * Copyright 2026 51 Degrees Mobile Experts Limited, Davidson House,
+ * Forbury Square, Reading, Berkshire, United Kingdom RG1 3EU.
+ *
+ * This Original Work is licensed under the European Union Public Licence
+ * (EUPL) v.1.2 and is subject to its terms as set out below.
+ *
+ * If a copy of the EUPL was not distributed with this file, You can obtain
+ * one at https://opensource.org/licenses/EUPL-1.2.
+ *
+ * The 'Compatible Licences' set out in the Appendix to the EUPL (as may be
+ * amended by the European Commission) shall be deemed incompatible for
+ * the purposes of the Work and the provisions of the compatibility
+ * clause in Article 5 of the EUPL shall not apply.
+ *
+ * If using the Work as, or as part of, a network application, by
+ * including the attribution notice(s) required under Article 5 of the EUPL
+ * in the end user terms of the application under an appropriate heading,
+ * such notice(s) shall fulfill the requirements of that article.
+ * ********************************************************************* */
+
+using FiftyOne.Did.Core.Data;
 using FiftyOne.Did.Core.FlowElements;
 using FiftyOne.Pipeline.CloudRequestEngine.Data;
 using FiftyOne.Pipeline.CloudRequestEngine.FlowElements;
@@ -11,10 +33,6 @@ using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-// ComponentMetaDataDefault lives in the FiftyOne.Did.Core project but
-// (inconsistently) under the FiftyOne.Did.OnPremise.Data namespace —
-// no dependency on a Did.OnPremise project, just the namespace.
-using FiftyOne.Did.OnPremise.Data;
 
 namespace FiftyOne.Did.Cloud.FlowElements
 {
