+++ FiftyOne.Pipeline.Core/Constants.cs
