+++ FiftyOne.Pipeline.Engines.FiftyOne/Data/FiftyOneAspectPropertyMetaDataDefault.cs
@@ -21,13 +21,12 @@
  * ********************************************************************* */
 
 using FiftyOne.Pipeline.Engines.Data;
-using FiftyOne.Pipeline.Engines.FiftyOne.Data;
 using FiftyOne.Pipeline.Engines.FlowElements;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace FiftyOne.Did.OnPremise.Data
+namespace FiftyOne.Pipeline.Engines.FiftyOne.Data
 {
     /// <summary>
     /// Data class that contains meta-data relating to a specific
