+++ FiftyOne.Pipeline.Engines.FiftyOne/Data/ValueMetaDataDefault.cs
@@ -20,10 +20,9 @@
  * such notice(s) shall fulfill the requirements of that article.
  * ********************************************************************* */
 
-using FiftyOne.Pipeline.Engines.FiftyOne.Data;
 using System;
 
-namespace FiftyOne.Did.OnPremise.Data
+namespace FiftyOne.Pipeline.Engines.FiftyOne.Data
 {
     /// <summary>
     /// Data class that contains meta-data relating to a specific
